+++ src/lua/client/ISUI/PermanentContextMenu.lua
@@ -4,8 +4,13 @@
 -- that can be found in the LICENSE file.
 --
 
+require 'PermanentRecipes'
+
 -- PermanentContextMenu is the game context menu extender.
 PermanentContextMenu = {}
+
+PermanentContextMenu.ghs = " <RGB:" .. getCore():getGoodHighlitedColor():getR() .. "," .. getCore():getGoodHighlitedColor():getG() .. "," .. getCore():getGoodHighlitedColor():getB() .. "> "
+PermanentContextMenu.bhs = " <RGB:" .. getCore():getBadHighlitedColor():getR() .. "," .. getCore():getBadHighlitedColor():getG() .. "," .. getCore():getBadHighlitedColor():getB() .. "> "
 
 -- doContextMenu adds the "ContextMenu_Moonshine" item to the context menu.
 PermanentContextMenu.doContextMenu = function(player, context, worldobjects, test)
@@ -26,6 +31,7 @@ PermanentContextMenu.doContextMenu = function(player, context, worldobjects, tes
     end
 
     local character = getSpecificPlayer(player)
+    local inventory = character:getInventory()
 
     local object = worldobjects[1];
     local square = object:getSquare()
@@ -53,8 +59,18 @@ PermanentContextMenu.doContextMenu = function(player, context, worldobjects, tes
         local distilMenu = ISContextMenu:getNew(context);
         context:addSubMenu(distilOption, distilMenu);
 
+        -- TODO: Iterate on PermanentRecipes.
         if SandboxVars.Permanent.AllowBrewingVanillaAlcohol then
-            -- TODO: Implement me.
+            local recipe = PermanentRecipes.MakeWhiskey;
+            local cookingSkill = character:getPerkLevel(Perks.Cooking);
+
+            if cookingSkill >= recipe.cookingSkill then
+                local option = distilMenu:addOption(getText("ContextMenu_MakeWhiskey"), worldobjects, PermanentContextMenu.OnBrew, character, object, recipe);
+
+                local toolTip = PermanentContextMenu.canBrew(option, character, recipe);
+                toolTip:setName(getText("ContextMenu_MakeWhiskey"));
+                toolTip:setTexture("Item_WhiskeyFull") -- TODO: get from recipe.results[0]
+            end
         end
 
         if SandboxVars.Permanent.AllowBrewingExclusiveAlcohol then
@@ -63,6 +79,51 @@ PermanentContextMenu.doContextMenu = function(player, context, worldobjects, tes
     end
 
     return true
+end
+
+PermanentContextMenu.canBrew = function(option, character, recipe)
+    local inventory = character:getInventory()
+
+    local tooltip = ISWorldObjectContextMenu.addToolTip();
+    option.toolTip = tooltip;
+    local result = true;
+    tooltip.description = getText("Tooltip_craft_Needs") .. ": <LINE>";
+
+    for itemCode, neededItemsCount in pairs(recipe.usedItems) do
+        local items = inventory:getAllType(itemCode)
+        local itemsCount = 0
+        if items then
+            for i=1, items:size() do
+                local itemToRemove = items:get(i-1);
+                if not PermanentRecipes.IsItemBlocked(character, itemToRemove) then
+                    itemsCount = itemsCount + 1;
+                end
+            end
+        end
+
+        local color = PermanentContextMenu.ghs;
+        if itemsCount < neededItemsCount then
+            color = PermanentContextMenu.bhs
+            result = false;
+        end
+
+        tooltip.description = tooltip.description .. color .. getItemNameFromFullType(itemCode) .. " " .. itemsCount .. "/" .. neededItemsCount .. " <LINE>";
+    end
+
+    if not result then
+        option.onSelect = nil;
+        option.notAvailable = true;
+    end
+
+    tooltip.description = " " .. tooltip.description .. " ";
+
+    return tooltip;
+end
+
+PermanentContextMenu.OnBrew = function(worldobjects, character, object, recipe)
+    if ISCampingMenu.walkToCampfire(character, object:getSquare()) then
+        ISTimedActionQueue.add(PermanentsBrewingAction:new(character, recipe, object));
+    end
 end
 
 Events.OnFillWorldObjectContextMenu.Add(PermanentContextMenu.doContextMenu);