+++ CHANGELOG.md
@@ -4,12 +4,17 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Fixed DoubleGreedyAxe Russian tooltip.
+- Fixed Sandbox_RU encoding.
+
 ### Changed
 - Changed all drinks icons.
 
 ## Added
 - Added Moonshine still building item (thanks for imopssi and Unfainthful).
 - Added Make Whiskey option on Moonshine still. 
+- Added Make Nicotine Overdose option on Moonshine still. 
 - Added 3d models (thanks for Unfainthful).
 - Added SlenderDoeSetWeight sandbox option.