+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v0.11.0] - 2023-10-11
 ### Fixed
 - Fixed DoubleGreedyAxe Russian tooltip.
 - Fixed Sandbox_RU encoding.
@@ -96,7 +98,8 @@ All notable changes to this project will be documented in this file.
 Slender Doe. 
 - Added EN and RU translations.
 
-[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/gorcon/rcon-cli/compare/v0.11.0...HEAD
+[v0.11.0]: https://github.com/gorcon/rcon-cli/compare/v0.10.0...v0.11.0
 [v0.10.0]: https://github.com/gorcon/rcon-cli/compare/v0.9.0...v0.10.0
 [v0.9.0]: https://github.com/gorcon/rcon-cli/compare/v0.8.0...v0.9.0
 [v0.8.0]: https://github.com/gorcon/rcon-cli/compare/v0.7.1...v0.8.0